gh-77: Fix user role delete

+++ custos-demo-gateway/src/new-service/store/modules/user.store.js
@@ -74,14 +74,18 @@ const actions = {
                 clientLevel: false
             });
 
-            await custosService.users.deleteRolesFromUser({
-                clientId,
-                roles: getters.getUser({
-                    clientId, username
-                }).realmRoles.filter(realmRole => realmRoles.indexOf(realmRole) < 0),
-                usernames: [username],
-                clientLevel: false
-            });
+            const realmRolesToBeDeleted = getters.getUser({
+                clientId, username
+            }).realmRoles.filter(realmRole => realmRoles.indexOf(realmRole) < 0);
+
+            if (realmRolesToBeDeleted.length > 0) {
+                await custosService.users.deleteRolesFromUser({
+                    clientId,
+                    roles: realmRolesToBeDeleted,
+                    username: username,
+                    clientLevel: false
+                });
+            }
         }
 
         if (clientRoles && clientRoles.length > 0) {
@@ -92,14 +96,18 @@ const actions = {
                 clientLevel: true
             });
 
-            await custosService.users.deleteRolesFromUser({
-                clientId,
-                roles: getters.getUser({
-                    clientId, username
-                }).clientRoles.filter(clientRole => clientRoles.indexOf(clientRole) < 0),
-                usernames: [username],
-                clientLevel: true
-            });
+            const clientRolesToBeDeleted = getters.getUser({
+                clientId, username
+            }).clientRoles.filter(clientRole => clientRoles.indexOf(clientRole) < 0);
+
+            if (clientRolesToBeDeleted.length) {
+                await custosService.users.deleteRolesFromUser({
+                    clientId,
+                    roles: clientRolesToBeDeleted,
+                    username: username,
+                    clientLevel: true
+                });
+            }
         }
 
         let updatedUser = await custosService.users.updateProfile({clientId, username, firstName, lastName, email});
